Programatically set approot at build time.
Because the approot will be different for each compute instance the
server runs on, programatically set it when the server is built then
installed. The domain name can be determined by querying the instances
metadata at 169.254.169.254.

Just prior to build, generate a header file containing macros defining
the hostname and port to use. Import this header file in the appropriate
Haskell files.