Reorganised handlers.
The problem with the way we had implemented our handlers is that
HTTP PUT and DELETE methods are not supported on all browsers.
This means that we need to simulate certain operations using
POST methods instead.

I wrote the following to Ben Lever in an email:

To make matters worse, you can't have any method by GET or POST
for forms in HTML (again invalidating the use of PUT).

PUT and DELETE do still have utility but mainly for web services
and libraries that interact with those web services (using some
kind of HTTP library). It's just browsers that are lagging behind
the game.

I changed the definition of the Foundation data type to:

/effects                       ListEffectsR            GET
/effects/#String/show          ShowEffectR             GET
/effects/#String/edit          EditEffectR             GET
/effects/#String/create        CreateEffectR           POST PUT
/effects/#String/update        UpdateEffectR           POST PUT
/effects/#String/delete        DeleteEffectR           POST DELETE
/effects/#String/run           RunEffectR              GET POST