Support upload of JPEG files rather than BMP.
Modify uploading of images to expect JPEGs rather than bitmaps for the
obvious bandwidth savings. However, as it proves to be a nightmare to
decode a JPEG image file, so as to access raw pixels, in Haskell, simply
call out to the "convert" application (imageMagik) to covert between
JPEGs and bitmaps. Using this approach, the effect code wrapper does
require modification.