Have image resource handlers compile/run effects.
Re-architect the image "preview" and "result" image resources to assume
that the image files do not exist. If the file does not exist, have the
handler compile and run the appropriate effect to generate it. Save the
resultant image so that is cached for the next time it is requested,
therefore avoiding re-compiling and re-running the effect.

Add preview images to the 'list', 'show' and 'edit' pages.

On the 'edit' page, change the "Save" button to "Save and Try" with the
behavior of saving the code, compiling it and displaying any compiler
error messages. If the compilation is successful the preview image is
simply displayed.

Add a "compiles" field to the 'Effect' entity that is set to True if the
effect's code compiles successfully, and False otherwise. Use this field
in conjunction with a new query prameter to the "/effects" resource,
"?compiles", to allow the specification of only listing those effects
which compile when set to "yes", otherwise all effects are listed.